refactor(server): move waddle-sfu livekit tests into a child module
livekit.rs was 1926 lines with over a thousand of them tests. Moving the
test module to livekit/tests.rs leaves the production file at 876 and
matches the repo's preference for small focused files. A child module
keeps access to the parent's private state (grant_locks, desired_grants,
issued) that several invariants assert on directly. No test changes; 82
still pass.
